Move vehicle sales and stock calculation into transportation calculations file
